refactor : app 폴더 비즈니스 로직, UI dmains 폴더로 분리
