+++ src/app/(app)/layout.tsx
@@ -1,4 +1,4 @@
-import { AppShell } from "../../components/layout/AppShell";
+import { AppShell } from "@/shared/ui/AppShell";
 
 export default function AppLayout({ children }: { children: React.ReactNode }) {
   return <AppShell>{children}</AppShell>;
