+++ src/domains/clips/utils/mapStoredClipDates.ts
@@ -1,0 +1,9 @@
+import { StoredClip } from "@/store/clipStore";
+import { Clip } from "@/types/clip";
+
+export const mapStoredClipDates = (clip: StoredClip): Clip => ({
+  ...clip,
+  createdAt: new Date(clip.createdAt),
+  updatedAt: clip.updatedAt ? new Date(clip.updatedAt) : undefined,
+  lastCopiedAt: clip.lastCopiedAt ? new Date(clip.lastCopiedAt) : null,
+});
