+++ src/domains/folders/types/folder.ts
@@ -1,0 +1,4 @@
+export interface FolderItem {
+  id: string;
+  name: string;
+}
